Creates A Public Constants Class for Client Use
It is helpful to expose certain constants for use by users of the
ClientAPI. Specifically in this case, the `MaxReadSize` value as it
limits how the user interacts with the connection. Exposing the constant
will allow the user to use it in the various "Read..." method calls
without having to dig into the source code to find the maximum or fail
and find it via trial and error.  This change creates a new, public
`ClientApiConstants` class that exposes a single `MaxReadSize` property
that is set as the value of `Consts.MaxReadSize`. It also revises all of
the previous `Consts.MaxReadSize` usages to use the new public version.
In looking at the usages of the other properties in the `Consts` class,
they did not seem as useful in normal usage of `ClientAPI`, so they were
not replicated in `ClientApiConstants`.  I left the bare value setting
of the `Consts.MaxReadSize` so the values could continue to be set in a
central location.